chat-widget/ui: Move relevant webpack configs to apps/
Change-type: patch
